+++ packages/veui/src/components/Overlay.vue
@@ -172,6 +172,7 @@ export default {
 
     focus () {
       this.overlayNode.toTop()
+      this.focusContext.toTop()
     },
 
     initFocus () {