+++ packages/veui/src/managers/alert.js
@@ -1,20 +1,16 @@
-import SpecialDialog from './SpecialDialog'
+import SimpleDialog from './SimpleDialog'
 import AlertBox from '../components/AlertBox'
 import Vue from 'vue'
 import { isFunction, noop, pick } from 'lodash'
 
-export class AlertManager extends SpecialDialog {
-  constructor () {
-    super(AlertBox)
-  }
-
+export class AlertManager extends SimpleDialog {
   createComponent (data) {
     const component = new Vue({
       render: h => {
         return h(
-          this.type,
+          AlertBox,
           {
-            props: pick(data, ['open', 'title', 'type', 'overlayClass']),
+            props: pick(data, ['open', 'title', 'type', 'overlayClass', 'beforeClose']),
             on: {
               ok: data.ok
             }
@@ -32,7 +28,7 @@ export class AlertManager extends SpecialDialog {
       let component = this.create({
         ...options,
         ok: () => {
-          Promise.resolve(ok()).then(result => {
+          Promise.resolve(ok()).then(() => {
             this.removeComponent(component)
             resolve()
           })