+++ packages/veui/src/utils/dom.js
@@ -211,31 +211,16 @@ export function focusAfter (elem) {
 }
 
 /**
- * 通过程序 focus 时手动添加 `.focus-visible` 类，弥补当前 polyfill 的不足。
- * 在不支持 `classList` 的浏览器下啥都不做，因为 polyfill 依赖了 `classList` polyfill。
- * 如果不引 `classList`，本来就不能工作，也就无需添加类。
- * 如果 `focus-visible` polyfill 修改了行为，也就不需要进行这样的处理了。
+ * 安全地 focus 一个元素
  *
  * @param {HTMLElement} elem
  */
 export function focus (elem) {
-  if (!elem) {
+  if (!elem || typeof elem.focus !== 'function') {
     return
   }
 
   elem.focus()
-  if (!elem.classList) {
-    return
-  }
-
-  let handler = () => {
-    elem.classList.remove('focus-visible')
-    elem.removeEventListener('blur', handler, false)
-    elem.removeEventListener('mouseleave', handler, false)
-  }
-  elem.addEventListener('blur', handler, false)
-  elem.addEventListener('mouseleave', handler, false)
-  elem.classList.add('focus-visible')
 }
 
 let transformKey